requirements.txt
annotated-types==0.7.0
black==24.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.7
et_xmlfile==2.0.0
flake8==7.1.1
google-ai-generativelanguage==0.6.1
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
httplib2==0.31.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy==1.11.2
mypy_extensions==1.1.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pydantic==2.11.10
pydantic_core==2.33.2
pyflakes==3.2.0
pyparsing==3.2.5
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.32
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
